requirements.txt
# --extra-index-url https://repo.dom_name/artifactory/api/pypi/proj/simple

allure-pytest==2.8.10
allure-python-commons==2.8.10
attrs==19.3.0
importlib-metadata==1.5.0
more-itertools==8.2.0
packaging==20.1
pluggy==0.13.1
py==1.8.1
pyparsing==2.4.6
pytest==5.3.5
selenium==3.141.0
six==1.14.0
urllib3==1.25.8
wcwidth==0.1.8
zipp==2.2.0
PyYAML==5.3
requests==2.23.0
psycopg2==2.8.4
rstr==2.2.6
pytest-xdist==1.31.0
pytest-dependency==0.5.1
rarfile
SQLAlchemy
jsondiff
pandas
jinja2
paramiko
dotted
jsonpath
backoff
faker
python-dateutil
